Added gpio controlled mouth and snow

+++ VGA_Graphics/Santa_Fight_System/vga16_graphics.h
@@ -36,7 +36,7 @@ enum colors {BLACK, DARK_GREEN, MED_GREEN, GREEN,
 // VGA primitives - usable in main
 void initVGA(void) ;
 void drawPixel(short x, short y, char color) ;
-void drawImg(uint8_t img[], uint8_t img_x, uint8_t img_y);
+void drawImg(uint8_t img[], int16_t img_x, int16_t img_y);
 void drawVLine(short x, short y, short h, char color) ;
 void drawHLine(short x, short y, short w, char color) ;
 void drawLine(short x0, short y0, short x1, short y1, char color) ;
